Moved enum class to common.py
